docs/pyproject.toml
[project]
name = "docs"
version = "0.1.0"
description = "flect docs"
authors = [
    { name = "Chaoying", email = "chaunceywe@gmail.com" }
]
dependencies = [
    "flect>=0.2.6",
    "pydantic-settings>=2.2.1",
]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/docs"]
